long->int

+++ src/main/java/com/fours/tolevelup/user/User.java
@@ -20,10 +20,10 @@ public class User {
 
     private String name;
     private String email;
-    private long level;
+    private int level;
 
     @Builder
-    public User(String id,String password,String name,String email,long level){
+    public User(String id,String password,String name,String email,int level){
         this.id = id;
         this.password = password;
         this.name = name;
